refactor(backend-db): rename FindTypeOrmServiceV2 module
