+++ packages/backend/libraries/backend-db/src/index.ts
@@ -2,7 +2,7 @@ import { FindQueryTypeOrmFromQueryBuilder } from './persistence/builders/typeorm
 import { DeleteTypeOrmQueryBuilderService } from './persistence/services/typeorm/DeleteTypeOrmQueryBuilderService';
 import { DeleteTypeOrmService } from './persistence/services/typeorm/DeleteTypeOrmService';
 import { FindTypeOrmQueryBuilderService } from './persistence/services/typeorm/FindTypeOrmQueryBuilderService';
-import { FindTypeOrmServiceV2 } from './persistence/services/typeorm/FindTypeOrmServiceV2';
+import { FindTypeOrmServiceV2 } from './persistence/services/typeorm/FindTypeOrmService';
 import { InsertTypeOrmService } from './persistence/services/typeorm/InsertTypeOrmService';
 import { InsertTypeOrmPostgresService } from './persistence/services/typeorm/postgres/InsertTypeOrmPostgresService';
 import { UpdateTypeOrmQueryBuilderService } from './persistence/services/typeorm/UpdateTypeOrmQueryBuilderService';
